# checkout-be/src/main/java/io/labs64/checkout/dto/PaginatedResponse.java
package io.labs64.checkout.dto;

import java.util.List;

import com.fasterxml.jackson.annotation.JsonInclude;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@Builder
public class PaginatedResponse<T> {
    final List<T> items;

}