# checkout-be/src/main/resources/openapi/openapi-checkout.yaml
openapi: 3.0.0
info:
  title: Checkout API
  version: 1.0.0
  description: REST API for handling the checkout process.
servers:
  - url: /api/v1
    description: Base path for Checkout API
...
paths:
...
  /checkout-intent:
    post:
      summary: Create a new checkout intent
      operationId: createCheckoutIntent
      tags:
        - CheckoutIntent
      description: |
        Creates a new checkout intent that represents a pending payment process.
        The checkout intent can include optional initial items, customer information,
        and metadata to customize the payment flow.
      requestBody:
        description: Payload to create a checkout intent.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutIntentCreateRequest'
      responses:
        '201':
          description: Checkout intent created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutIntent'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "VALIDATION_ERROR"
                message: "Currency is invalid. Allowed currencies: USD, EUR"
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    get:
      summary: Returns a list of checkout intents
      operationId: listCheckoutIntents
      tags:
        - CheckoutIntent
      description: |
        Retrieves all existing checkout intents.
        This endpoint is typically used for administrative or reporting purposes
        to review created payment attempts.
      x-spring-paginated: true
      parameters:
        - name: query
          in: query
          description: Free-text search term to filter checkout intents.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of checkout intents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutIntentPage'
  /checkout-intent/{id}:
    get:
      summary: Retrieve a checkout intent by its ID
      operationId: getCheckoutIntent
      tags:
        - CheckoutIntent
      description: |
        Retrieves the details of a specific checkout intent using its unique identifier.
        This can be used to check the current status or metadata of a pending or completed payment.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the checkout intent to retrieve.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Checkout intent found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutIntent'
        '404':
          description: Checkout intent not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "NOT_FOUND"
                message: "CheckoutIntent with ID '550e8400-e29b-41d4-a716-446655440000' was not found."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    patch:
      summary: Partially update a checkout intent
      operationId: updateCheckoutIntent
      tags:
        - CheckoutIntent
      description: |
        Partially updates an existing checkout intent.
        This can be used to modify fields like status, payment method, 
        or billing details without replacing the entire object.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the checkout intent to retrieve.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        description: Fields to update in the checkout intent.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutIntentUpdateRequest'
      responses:
        '200':
          description: Checkout intent updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutIntent'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "VALIDATION_ERROR"
                message: "Currency is invalid. Allowed currencies: USD, EUR"
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        '404':
          description: Checkout intent not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "NOT_FOUND"
                message: "CheckoutIntent with ID '550e8400-e29b-41d4-a716-446655440000' was not found."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"

  # BILLING INFO
  /billing-info:
    post:
      summary: Create a new billing info
      operationId: createBillingInfo
      tags:
        - BillingInfo
      description: |
        Creates a new billing info record for a customer. 
        This info can later be linked to a CheckoutIntent for processing payments.
      requestBody:
        description: Payload to create a billing info.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingInfoCreateRequest'
      responses:
        '201':
          description: Billing info created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingInfo'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "VALIDATION_ERROR"
                message: "Email is invalid."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    get:
      summary: Returns a list of billing infos
      operationId: listBillingInfos
      tags:
        - BillingInfo
      description: Retrieves a paginated list of billing info records, optionally filtered by a free-text search term.
      x-spring-paginated: true
      parameters:
        - name: query
          in: query
          description: Free-text search term to filter billing info.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of billing infos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingInfoPage'
  /billing-info/{id}:
    get:
      summary: Retrieve a billing info by its ID
      operationId: getBillingInfo
      tags:
        - BillingInfo
      description: Retrieves the details of a specific billing info by its UUID.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the billing info to retrieve.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Billing info found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingInfo'
        '404':
          description: Billing info not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "NOT_FOUND"
                message: "BillingInfo with ID '550e8400-e29b-41d4-a716-446655440000' was not found."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    patch:
      summary: Partially update a billing info
      operationId: updateBillingInfo
      tags:
        - BillingInfo
      description: |
        Partially updates a billing info record. Only non-critical fields can be updated if the billing 
        info is already linked to a CheckoutIntent.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the billing info to update.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        description: Fields to update in the billing info.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillingInfoUpdateRequest'
      responses:
        '200':
          description: Billing info updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillingInfo'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "VALIDATION_ERROR"
                message: "email is invalid."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        '404':
          description: Billing info not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "NOT_FOUND"
                message: "BillingInfo with ID '550e8400-e29b-41d4-a716-446655440000' was not found."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    delete:
      summary: Delete a billing info
      operationId: deleteBillingInfo
      tags:
        - BillingInfo
      description: |
        Deletes a billing info record if it is not currently linked to any CheckoutIntent. 
        Deletion is forbidden for billing infos that are in use.
      parameters:
        - name: id
          in: path
          required: true
          description: UUID of the billing info to delete.
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Billing info deleted successfully.
        '404':
          description: Billing info not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "NOT_FOUND"
                message: "BillingInfo with ID '550e8400-e29b-41d4-a716-446655440000' was not found."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        '409':
          description: BillingInfo is linked to CheckoutIntents and cannot be deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: "CONFLICT"
                message: "BillingInfo is linked to CheckoutIntents and cannot be deleted."
                timestamp: "2025-08-19T12:34:56Z"
                traceId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"

  # SHIPPING INFO
...
components:
  schemas:
    CheckoutIntentStatus:
      type: string
      description: Represents the current state of a checkout intent.
      enum:
        - CREATED     # The checkout intent has been created but not yet processed.
        - PENDING     # The checkout is awaiting payment confirmation or further action.
        - FAILED      # The payment attempt failed or was declined.
        - CANCELED    # The checkout intent was canceled before completion.
        - COMPLETED   # The payment was successfully completed.
    CheckoutIntentBase:
      type: object
      properties:
        status:
          allOf:
            - $ref: '#/components/schemas/CheckoutIntentStatus'
            - readOnly: true
        amount:
          type: integer
          format: int64
          description: Total amount to be charged (minor currency units).
          minimum: 0
          example: 14999
        currency:
          allOf:
            - $ref: '#/components/schemas/Currency'
          x-extra-annotation: "@io.labs64.checkout.validator.ValidCurrency"
        paymentMethod:
          type: string
          description: Unique identifier of the payment method.
          example: "STRIPE"
        billingInfoId:
          type: string
          format: uuid
          description: Unique identifier of the billing info.
          example: "550e8400-e29b-41d4-a716-446655440000"
        purchaseOrderId:
          type: string
          format: uuid
          description: Unique identifier of the purchase order.
          example: "550e8400-e29b-41d4-a716-446655440000"
        extra:
          $ref: '#/components/schemas/Extra'
    CheckoutIntent:
      type: object
      description: Represents a payment attempt.
      allOf:
        - $ref: '#/components/schemas/CheckoutIntentBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: Unique identifier of the checkout intent.
              readOnly: true
              example: "550e8400-e29b-41d4-a716-446655440000"
            createdAt:
              type: string
              format: date-time
              description: Timestamp when the checkout intent was created.
              readOnly: true
              example: "2025-09-11T12:34:56Z"
            updatedAt:
              type: string
              format: date-time
              description: Timestamp when the checkout intent was last updated.
              readOnly: true
              example: "2025-09-11T13:20:00Z"
            closedAt:
              type: string
              format: date-time
              description: Timestamp when the checkout intent reached a final status (COMPLETED, CANCELED, or FAILED).
              readOnly: true
              example: "2025-09-11T14:00:00Z"
    CheckoutIntentCreateRequest:
      allOf:
        - $ref: '#/components/schemas/CheckoutIntentBase'
        - type: object
          description: Schema used when creating a new checkout intent. Requires mandatory fields such as amount and currency.
          required:
            - amount
            - currency
            - paymentMethod
    CheckoutIntentUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/CheckoutIntentBase'
        - type: object
          description: Schema used when partially updating an existing checkout intent. Only the provided fields will be modified.
    CheckoutIntentPage:
      type: object
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/CheckoutIntent'
              default: [ ]
          required:
            - items
    BillingInfoBase:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          description: Full name of the person responsible for billing.
          maxLength: 255
          example: "John Doe"
        email:
          type: string
          format: email
          description: Email address for billing purposes.
          maxLength: 255
          example: "john.doe@example.com"
        phone:
          type: string
          description: Phone number for billing purposes.
          maxLength: 50
          example: "+1234567890"
        city:
          type: string
          description: City of the billing address.
          maxLength: 100
          example: "New York"
        country:
          type: string
          description: Country code of the billing address in ISO 3166-1 alpha-2 format.
          maxLength: 2
          example: "US"
        address1:
          type: string
          description: Primary street address.
          maxLength: 255
          example: "123 Main St"
        address2:
          type: string
          description: Secondary street address (optional).
          maxLength: 255
          example: "Apt 4B"
        postalCode:
          type: string
          description: ZIP or postal code of the billing address.
          maxLength: 20
          example: "10001"
        state:
          type: string
          description: State or region of the billing address.
          maxLength: 50
          example: "NY"
        extra:
          allOf:
            - $ref: '#/components/schemas/Extra'
          x-extra-annotation: "@io.labs64.checkout.validator.ValidExtra"
    BillingInfo:
      type: object
      description: Billing details associated with this payment attempt.
      allOf:
        - $ref: '#/components/schemas/BillingInfoBase'
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: Unique identifier of the billing info.
              readOnly: true
              example: "550e8400-e29b-41d4-a716-446655440000"
            createdAt:
              type: string
              format: date-time
              description: Timestamp when the billing info was created.
              readOnly: true
              example: "2025-09-11T12:34:56Z"
            updatedAt:
              type: string
              format: date-time
              description: Timestamp when the billing info was last updated.
              readOnly: true
              example: "2025-09-11T13:20:00Z"
            confirmedAt:
              type: string
              format: date-time
              description: Timestamp when the billing info was confirmed (user consent recorded).
              readOnly: true
              example: "2025-09-11T13:20:00Z"
    BillingInfoCreateRequest:
      allOf:
        - $ref: '#/components/schemas/BillingInfoBase'
        - type: object
          description: Schema used when creating a new billing info.
          required:
            - name
          properties:
            confirm:
              type: boolean
              enum: [ true ]
              description: "Explicit user consent to store billing info. Must be true."
    BillingInfoUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/BillingInfoBase'
        - type: object
          description: Schema used when partially updating an existing billing info. Only the provided fields will be modified.
    BillingInfoPage:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/BillingInfo'
              default: [ ]
          required:
            - items
...
    Currency:
      type: string
      description: ISO-4217 currency code (3 uppercase letters).
      pattern: '^[A-Z]{3}$'
      example: USD
    Extra:
      type: object
      description: Custom user-defined fields.
      additionalProperties: true
      example:
        source: "web_portal"
        promoCode: "SUMMER2025"
    ErrorCode:
      type: string
      description: Unique error code identifying the type of error.
      enum:
        - NOT_FOUND         # The requested resource was not found.
        - INTERNAL_ERROR    # An unexpected server-side error occurred.
        - VALIDATION_ERROR  # The request payload failed validation rules.
        - CONFLICT          # Resource conflict (e.g., FK restriction, duplicate, etc.)
        - CONSENT_REQUIRED  # Consent required
        - MISSING_TENANT_ID # Missing header "X-Tenant-Id"
      example: "NOT_FOUND"
    ErrorResponse:
      type: object
      description: Standard error response returned by the Checkout API.
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Human-readable message describing the error.
          example: "Checkout with ID '550e8400-e29b-41d4-a716-446655440000' was not found."
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred.
          example: "2025-08-19T12:34:56Z"
        traceId:
          type: string
          description: Optional trace ID for correlating logs or debugging.
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
    PaginatedResponse:
      type: object
      description: Standard pagination envelope used by list endpoints. Contains page cursors and total counters.
      properties:
        page:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 100
        totalItems:
          type: integer
          format: int64
          example: 1000
        totalPages:
          type: integer
          example: 10
        hasPrev:
          type: boolean
        hasNext:
          type: boolean
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
        - hasPrev
        - hasNext